Enhancements
- Project variable separated in another file
- iterators and function included in the file to improve efficiency, and avoid code repetition
