+++ projects.js
@@ -1,0 +1,38 @@
+export const projects = [
+  {
+    title: 'Budget App',
+    tech: ['Rails', 'Ruby', 'Tailwind'],
+    image: './images/budget.png',
+    desc_short: 'The <strong>Budget Application</strong> is a mobile web-based tool designed to facilitate easy management of personal finances. The application provides a comprehensive list of transactions organized by category, enabling users to track their spending habits and identify areas of improvement',
+    desc_long: 'The <strong>Budget Application</strong> is a mobile web-based tool designed to facilitate easy management of personal finances. The application provides a comprehensive list of transactions organized by category, enabling users to track their spending habits and identify areas of improvement. Developed using the Ruby on Rails framework, both for the front-end and back-end, and styled with the Tailwind framework, the Budget Application delivers a seamless user experience while maintaining a sleek and professional design.',
+    demo: 'https://small-wave-7734.fly.dev/',
+    src: 'https://github.com/HansZizold/budget_app',
+  },
+  {
+    title: 'Blog App',
+    tech: ['Rails', 'Ruby', 'CSS3'],
+    image: './images/blog.png',
+    desc_short: 'The Blog App allows for efficient management of blog posts, comments, and likes, empowering users to interact and engage with content seamlessly. With its user-friendly interface and intuitive design, Blog App makes it easy',
+    desc_long: "Blog App is fully functional website. The application allows for efficient management of blog posts, comments, and likes, empowering users to interact and engage with content seamlessly. With its user-friendly interface and intuitive design, Blog App makes it easy to create and publish new posts, moderate comments, and monitor engagement. Whether you're a blogger looking to expand your reach or a business seeking to enhance your content marketing strategy, Blog App is an ideal solution that can help you achieve your goals",
+    demo: 'https://cold-brook-8415.fly.dev/',
+    src: 'https://github.com/HansZizold/blogapp',
+  },
+  {
+    title: 'Hyrule Compendium',
+    tech: ['React', 'Redux', 'CSS3'],
+    image: './images/hyrule.png',
+    desc_short: "Hyrule Compendium is a comprehensive web application that showcases a vast array of elements from the 'Breath of the Wild' Zelda game. With an intuitive user interface, Hyrule Compendium allows you easily browse and discover",
+    desc_long: "Hyrule Compendium is a comprehensive web application that showcases a vast array of elements from the popular 'Breath of the Wild' Zelda game. With an intuitive user interface, Hyrule Compendium allows you to easily browse and discover your favorite characters, weapons, items, and locations within the game. Whether you're a seasoned Zelda fan or a new player exploring the world of Hyrule for the first time, Hyrule Compendium provides a wealth of information and resources to enhance your gaming experience",
+    demo: 'https://hyrule-compendium-hpz.netlify.app/',
+    src: 'https://github.com/HansZizold/Hyrule_Compendium',
+  },
+  {
+    title: 'Space Traveler Hub',
+    tech: ['React', 'Redux', 'CSS3'],
+    image: './images/rocket.png',
+    desc_short: 'Welcome to Space Travelers Hub, a comprehensive web application that allows you to explore a variety of rockets and missions. Our platform offers users the ability to reserve and join exciting missions, enabling them to discover...',
+    desc_long: 'Welcome to Space Travelers Hub, a comprehensive web application that allows you to explore a variety of rockets and missions. Our platform offers users the ability to reserve and join exciting missions, enabling them to discover the wonders of space travel like never before. With a user-friendly interface and cutting-edge features, Space Travelers Hub is the ultimate destination for those looking to embark on an adventure beyond our planet.',
+    demo: 'https://space-travelers-hd.netlify.app/',
+    src: 'https://github.com/HansZizold/Space-Travelers-Hub',
+  }
+];
